requirements.txt
django
djangorestframework
pyownet
markdown
